FIX: IniFileParser now properly replaces multiple variables per entry STYLE: Removed some compiler warnings
git-svn-id: https://svn.rwth-aachen.de/repos/vrgroup-svn/projects/Vista/trunk@33761 33c38b33-09d2-4561-8d4d-1bfa2f58f885

+++ VistaCoreLibs/VistaDeviceDrivers/VistaOpenALRecordDriver/VistaOpenALRecordDriver.cpp
@@ -571,7 +571,7 @@ bool VistaOpenALRecordDriver::DoSensorUpdate(VistaType::microtime nTs)
 	if( nToGo == -1 )
 		nToGo = samples; // normalize to samples
 
-	nToGo -= m_pCapture->captureTo( &pM->m_vecMeasures[(samples - nToGo)*numBytesPerSample], nToGo );
+	nToGo -= (int)m_pCapture->captureTo( &pM->m_vecMeasures[(samples - nToGo)*numBytesPerSample], nToGo );
 
 	if( nToGo == 0 )
 	{
